requirements.txt
# data science
numpy
pandas

# Uncomment if you use sklearn
# scikit-learn

# Install the correct TensorFlow version
# Uncomment if you use TensorFlow
tensorflow == 2.16.1

# Google Cloud
google-cloud-storage

python-dotenv

# API
fastapi
uvicorn

#Images
pillow
